+++ tt_text.h
@@ -14,6 +14,8 @@ int tt_text_to_mhead (const char *text, int quiet, char *buttons, size_t buttons
 
 int tt_text_to_satsq (const char *text, int quiet, char *buttons, size_t buttonsiz);
 
+int tt_text_to_ascii2d (const char *text, int quiet, char *buttons);
+
 
 /* Decode DTMF to normal human readable form. */
 
@@ -26,6 +28,9 @@ int tt_call10_to_text (const char *buttons, int quiet, char *text);
 int tt_mhead_to_text (const char *buttons, int quiet, char *text, size_t textsiz);
 
 int tt_satsq_to_text (const char *buttons, int quiet, char *text);
+
+int tt_ascii2d_to_text (const char *buttons, int quiet, char *text);
+
 
 
 /* end tt_text.h */